M3-B/C: Task 관리(원장 동기·링 커서·액션) + 시나리오 관리(러너·JSON/CSV 입출력·편집기); 통합 정리
- ledger/sync: 링 커서 sqlite 저장, 오버플로 감지, 에코 상관, Lost 재승격
- ledger/ops·routes: AUTO 모드 취소 409, stats, mark-failed, DELETE
- scenario: Runner(start/pause/resume/stop), 순수 플래너, io(JSON/CSV), validate, 실행 이력
- 프론트: components/taskmgr, components/scenario, lib/task/state, lib/scenario
- 리드: 워크스페이스 clippy -D warnings 0건, rustfmt 적용, Step PartialEq 파생, 데모로 PICK·시나리오 2스텝 완료 확인

+++ apps/gr-console/src/measure/mod.rs
@@ -68,7 +68,7 @@ impl MeasureStore {
             return Ok(0);
         }
         let layout = plc.layout("MEASLOG_HIST").ok_or_else(|| ApiError::Internal("MEASLOG_HIST not in contract".into()))?;
-        let capacity = layout.range_of("Entry").map(|(lo, hi)| (hi - lo) / entry_size(layout).max(1)).unwrap_or(200) as u32;
+        let capacity = layout.range_of("Entry").map(|(lo, hi)| (hi - lo) / entry_size(layout).max(1)).unwrap_or(200);
         let missing = (total - have).min(capacity);
         let mut n = 0;
         for k in 0..missing {
